Move instantiateImagecodec to painting/binding
