+++ packages/flutter/lib/src/painting/binding.dart
@@ -9,6 +9,8 @@ import 'package:flutter/services.dart' show ServicesBinding;
 
 import 'image_cache.dart';
 
+const double _kDefaultDecodedCacheRatioCap = 25.0;
+
 /// Binding for the painting library.
 ///
 /// Hooks into the cache eviction logic to clear the image cache.
@@ -46,9 +48,35 @@ abstract class PaintingBinding extends BindingBase with ServicesBinding {
   @protected
   ImageCache createImageCache() => ImageCache();
 
+  /// The maximum multiple of the compressed image size used when caching an
+  /// animated image.
+  ///
+  /// By default individual frames of animated images are cached into memory to
+  /// avoid using CPU to re-decode them for every loop in the animation. This
+  /// behavior will result in out-of-memory crashes when decoding large
+  /// (or large numbers of) animated images. Set this value to limit how much
+  /// memory each animated image is allowed to use to cache decoded frames
+  /// compared to its compressed size. For example, setting this to `2.0` means
+  /// that a 400KB GIF would be allowed at most to use 800KB of memory caching
+  /// unessential decoded frames. A setting of `1.0` or less disables all caching
+  /// of unessential decoded frames. See [_kDefaultDecodedCacheRatioCap] for the
+  /// default value.
+  double get decodedCacheRatioCap => _kDecodedCacheRatioCap;
+  double _kDecodedCacheRatioCap = _kDefaultDecodedCacheRatioCap;
+  /// Changes the maximum multiple of compressed image size used when caching an
+  /// animated image.
+  ///
+  /// Changing this value only affects new images, not images that have already
+  /// been decoded.
+  set decodedCacheRatioCap(double value) {
+    assert (value != null);
+    assert (value >= 0.0);
+    _kDecodedCacheRatioCap = value;
+  }
+
   /// Calls through to [dart:ui] with [decodedCacheRatioCap] from [ImageCache].
   Future<ui.Codec> instantiateImageCodec(Uint8List list) {
-    return ui.instantiateImageCodec(list, decodedCacheRatioCap: _imageCache.decodedCacheRatioCap);
+    return ui.instantiateImageCodec(list, decodedCacheRatioCap: decodedCacheRatioCap);
   }
 
   @override
